git-svn-id: file:///Users/kraman/Documents/personal/doc/RTZen/branches/poa-and-policies@1888 bfff1265-3dd4-0310-90a8-ce2828190deb

+++ packages/src/edu/uci/ece/zen/orb/protocol/IOP/ServiceContextHelper.java
@@ -1,16 +1,17 @@
 package edu.uci.ece.zen.orb.protocol.IOP;
 
 import edu.uci.ece.zen.utils.FString;
+import edu.uci.ece.zen.utils.ZenProperties;
 
 /**
  * Helper class for : ServiceContext
- * 
+ *
  * @author OpenORB Compiler
  */
 public class ServiceContextHelper {
     /**
      * Read ServiceContext from a marshalled stream
-     * 
+     *
      * @param istream
      *            the input stream
      * @return the readed ServiceContext value
@@ -18,13 +19,10 @@ public class ServiceContextHelper {
     public static FString read(org.omg.CORBA.portable.InputStream istream,
             FString fs) {
         //ServiceContext new_one = ServiceContext.instance(pos);
-
         int context_id = org.omg.IOP.ServiceIdHelper.read(istream);
-        fs.append((long) context_id);
-
+        fs.append((int) context_id);
         int context_data_length = istream.read_ulong();
-        fs.append((long) context_data_length);
-
+        fs.append((int) context_data_length);
         //istream.read_octet_array(new_one.context_data, 0,
         // new_one.context_data_length);
         fs.read(istream, context_data_length);
@@ -34,7 +32,7 @@ public class ServiceContextHelper {
 
     /**
      * Write ServiceContext into a marshalled stream
-     * 
+     *
      * @param ostream
      *            the output stream
      * @param value
